+++ R/utils-depigner.R
@@ -1,0 +1,11 @@
+# The following block is used by depigner to automatically mange
+# roxygen namespace tags to add usethis' User Interface functions to
+# the package's NAMESPACE. Modify with care!
+## depigner namespace: start
+#' @importFrom usethis ui_line ui_todo ui_done ui_todo ui_oops ui_info
+#' @importFrom usethis ui_code_block
+#' @importFrom usethis ui_stop ui_warn
+#' @importFrom usethis ui_yeah ui_nope
+#' @importFrom usethis ui_field ui_value ui_path ui_code
+## depigner namespace: end
+NULL